+++ subband.h
@@ -1,3 +1,3 @@
 void create_syn_filter (double filter[64][SBLIMIT]);
-int SubBandSynthesis (double *bandPtr, int channel, long *samples);
-int SubBandSynthesis_ml (double *bandPtr, int channel, long *samples);
+int SubBandSynthesis (double *bandPtr, int channel, int32_t *samples);
+int SubBandSynthesis_ml (double *bandPtr, int channel, int32_t *samples);
